fixes from @braingram comments
