+++ CHANGES.rst
@@ -1,7 +1,24 @@
 1.15.1 (unreleased)
 ===================
 
--
+cube_build
+----------
+
+- Removed direct setting of the ``self.skip`` attribute from within the step
+  itself. [#8600]
+
+stpipe
+------
+
+- Removed setting of the `self.skip` attribute in the `record_step_status()` function;
+  added a `query_step_status()` function to use as an alternative to checking 
+  `self.skip`. [#8600]
+
+tweakreg
+--------
+
+- Removed direct setting of the ``self.skip`` attribute from within the step
+  itself. [#8600]
 
 1.15.0 (2024-06-26)
 ===================
@@ -77,12 +94,6 @@ combine_1d
 ----------
 
 - Fix weights for combining errors from 1D spectra. [#8520]
-
-cube_build
-----------
-
-- Removed direct setting of the ``self.skip`` attribute from within the step
-  itself. [#8600]
 
 dark_current
 ------------
@@ -391,13 +402,6 @@ saturation
 
 - Adds a check for saturation bias in group 2 for IRS2 mode nframes > 1. [#8593]
 
-stpipe
-------
-
-- Removed setting of the `self.skip` attribute in the `record_step_status()` function,
-  added a `query_step_status()` function to use as an alternative to checking 
-  `self.skip`. [#8600]
-
 tweakreg
 --------
 
@@ -422,9 +426,6 @@ tweakreg
   that ``separation`` > ``sqrt(2) * tolerance`` that will now log an error
   message and skip ``tweakreg`` step when this condition is not satisfied and
   source confusion is possible during catalog matching. [#8476]
-
-- Removed direct setting of the ``self.skip`` attribute from within the step
-  itself. [#8600]
 
 wavecorr
 --------
