+++ jwst/stpipe/__init__.py
@@ -1,5 +1,5 @@
 from .core import JwstStep as Step, JwstPipeline as Pipeline
-from .core import record_step_status, query_step_status
+from .utilities import record_step_status, query_step_status
 
 
 __all__ = ['Step', 'Pipeline', 'record_step_status', 'query_step_status']
